dev commit

+++ ost/s1/batch_burst.py
@@ -26,7 +26,7 @@ def burst_to_ard_batch(
         processing_dir,
         ard_parameters,
         data_mount='/eodata',
-        max_workers=os.cpu_count()/2
+        max_workers=int(os.cpu_count()/2)
 ):
     '''Handles the batch processing of a OST complinat burst inventory file
 
@@ -39,7 +39,7 @@ def burst_to_ard_batch(
 
     '''
     if max_workers > os.cpu_count()/2:
-        max_workers = os.cpu_count()/2
+        max_workers = int(os.cpu_count()/2)
     executor_type = 'concurrent_processes'
     executor = Executor(executor=executor_type,
                         max_workers=max_workers
